Add pretty printers for FTL time types for use in gtest assertions
This teaches gtest to print out the numeric value of Time{Point,Delta}
instead of printing out the memory value of the type in hex. To avoid
bloating every binary that wants to use FTL with a bunch of stream
string formatting code, the printers are defined in a separate library
annotated with testonly = true that must be linked in to test targets
that contain gtest assertions about time values. The symbol for
printing Time{Point,Delta} is declared in the same header file as the
actual type to avoid ODR violations with multiple different
definitions of the printers.

In practice, what this means is that any test target that contains a
gtest assertion about these time values will fail to link if it fails
to link in the printer target as gtest will pick up on the declaration
in the header file for PrintTo.

Change-Id: I256b9ecdd98831edf99a77aa9cc7a0a14c31235f