Port weak_ptr and string_view to Windows
FTL needs to run on Windows in order to port Flutter's sky_shell to
Windows. We need sky_shell on Windows to run widget tests on the host.

With this FTL now compiles on Windows and all unittests pass. In a
future change, we will still have to move the Windows API functions
from their ANSI to their wide versions.

Change-Id: I5891ee10af339a86ad9e369c8f8b55270231bd9e