Restore some filename context to logging operations
In the case of WARN/ERROR/FATAL, the full path to the file will be
displayed, without the leading dots. In the case of INFO or verbose,
the filename will be displayed without the path. The theory is that an
error can happen in a filename you aren't familiar with, whereas INFO
will mostly be interesting to the developer of a particular file.

Change-Id: I3816c0694955fdcc793be465bf18ee0c55370afd